feat: Data Layer 요소 구현
- DataRepository 구현
- DataSource 프로토콜 구현
- Repository 프로토콜 파일 분리
